requirements.txt
pyTelegramBotAPI==4.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
anthropic==0.39.0
flask==3.0.3

# Dev / тесты — не ставятся в проде, но нужны локально
pytest==8.3.3
